refactor(terminal): extract event listeners into dedicated module
- Create src/store/terminalStoreListeners.ts with setupTerminalStoreListeners and cleanupTerminalStoreListeners
- Move normalizeCrashType, activity buffer, and 15 unsubscribe variables to new module
- Add re-exports from terminalStore.ts for backward compatibility
- Promote clearTerminalRestartGuard import to top-level in terminalStore.ts
- Remove listener-only imports from terminalStore.ts (AgentStateChangePayload, TerminalActivityPayload, TerminalStatusPayload, useResourceMonitoringStore, SCROLLBACK_BACKGROUND)
